Organised the structure of the compents folder
