+++ src/components/movies/movieCard/index.jsx
@@ -14,7 +14,7 @@ import Grid from "@mui/material/Grid";
 import img from '../../images/film-poster-placeholder.png'
 import { Link } from "react-router-dom";
 import Avatar from "@mui/material/Avatar";
-import { MoviesContext } from "../../contexts/moviesContext";
+import { MoviesContext } from "../../../contexts/moviesContext";
 
 const styles = {
   card: { maxWidth: 345 },
