+++ src/components/movies/movieReviews/index.jsx
@@ -7,8 +7,8 @@ import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
 import Paper from "@mui/material/Paper";
 import { Link } from "react-router-dom";
-import { getMovieReviews } from "../../api/tmdb-api";
-import { excerpt } from "../../util";
+import { getMovieReviews } from "../../../api/tmdb-api";
+import { excerpt } from "../../../util";
 
 const styles = {
   table: {
