+++ src/components/movies/templateMoviePage/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MovieHeader from "../headerMovie";
+import MovieHeader from "../movies/headerMovie";
 import Grid from "@mui/material/Grid";
 import ImageList from "@mui/material/ImageList";
 import ImageListItem from "@mui/material/ImageListItem";
