+++ src/pages/actorDetailsPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import ActorDetails from "../components/actorDetails";
+import ActorDetails from "../components/actors/actorDetails";
 import PageTemplate from "../components/templateActorPage";
 import { getActor } from '../api/tmdb-api'
 import { useQuery } from "react-query";
