+++ src/pages/homePage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PageTemplate from "../components/templateMovieListPage";
+import PageTemplate from "../components/movies/templateMovieListPage";
 import { useQuery } from "react-query";
 import Spinner from "../components/spinner";
 import { getMovies } from "../api/tmdb-api";
@@ -10,7 +10,7 @@ import MovieFilterUI, {
   titleFilter,
   genreFilter,
   voteFilter,
-} from "../components/movieFilterUI";
+} from "../components/movies/movieFilterUI";
 
 const titleFiltering = {
   name: "title",
