+++ src/pages/movieDetailsPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import MovieDetails from "../components/movieDetails";
+import MovieDetails from "../components/movies/movieDetails";
 import PageTemplate from "../components/templateMoviePage";
 import { getMovie } from '../api/tmdb-api'
 import { useQuery } from "react-query";
