+++ src/pages/movieReviewPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
 import PageTemplate from "../components/templateMoviePage";
-import MovieReview from "../components/movieReview";
+import MovieReview from "../components/movies/movieReview";
 
 const MovieReviewPage = (props) => {
   const { state : {movie, review } } = useLocation()
