+++ src/pages/similarMoviesPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useQuery } from "react-query";
 import { useParams } from "react-router-dom";
-import PageTemplate from '../components/templateMovieListPage'
+import PageTemplate from '../components/movies/templateMovieListPage'
 import { getSimilarMovies } from "../api/tmdb-api";
 import AddToPlaylistIcon from '../components/cardIcons/addToPlaylist';
 import Spinner from "../components/spinner";
@@ -10,7 +10,7 @@ import useFiltering from "../hooks/useFiltering";
 import MovieFilterUI, {
   titleFilter,
   genreFilter,
-} from "../components/movieFilterUI";
+} from "../components/movies/movieFilterUI";
 
 const titleFiltering = {
   name: "title",
